Moved classes to the new module
