+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/settings/FormMetadataSettingsTest.kt
@@ -8,7 +8,6 @@ import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
 import org.odk.collect.android.R
 import org.odk.collect.android.injection.config.AppDependencyModule
-import org.odk.collect.android.metadata.InstallIDProvider
 import org.odk.collect.android.support.pages.MainMenuPage
 import org.odk.collect.android.support.pages.ProjectSettingsPage
 import org.odk.collect.android.support.pages.SaveOrIgnoreDialog
@@ -16,7 +15,8 @@ import org.odk.collect.android.support.pages.UserAndDeviceIdentitySettingsPage
 import org.odk.collect.android.support.rules.CollectTestRule
 import org.odk.collect.android.support.rules.ResetStateRule
 import org.odk.collect.android.support.rules.TestRuleChain
-import org.odk.collect.android.utilities.DeviceDetailsProvider
+import org.odk.collect.metadata.DeviceDetailsProvider
+import org.odk.collect.metadata.InstallIDProvider
 
 @RunWith(AndroidJUnit4::class)
 class FormMetadataSettingsTest {
