+++ collect_app/src/test/java/org/odk/collect/android/preferences/screens/FormMetadataPreferencesFragmentTest.java
@@ -5,7 +5,7 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
-import static org.odk.collect.android.logic.PropertyManager.PROPMGR_DEVICE_ID;
+import static org.odk.collect.metadata.PropertyManager.PROPMGR_DEVICE_ID;
 import static org.odk.collect.settings.keys.ProjectKeys.KEY_METADATA_PHONENUMBER;
 
 import android.app.Activity;
@@ -22,10 +22,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.odk.collect.android.R;
 import org.odk.collect.android.injection.config.AppDependencyModule;
-import org.odk.collect.android.metadata.InstallIDProvider;
 import org.odk.collect.android.support.CollectHelpers;
-import org.odk.collect.android.utilities.DeviceDetailsProvider;
 import org.odk.collect.fragmentstest.FragmentScenarioLauncherRule;
+import org.odk.collect.metadata.DeviceDetailsProvider;
+import org.odk.collect.metadata.InstallIDProvider;
 import org.odk.collect.permissions.ContextCompatPermissionChecker;
 import org.odk.collect.permissions.PermissionListener;
 import org.odk.collect.permissions.PermissionsChecker;
