+++ collect_app/src/test/java/org/odk/collect/android/utilities/WebCredentialsUtilsTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import org.junit.Test;
-import org.odk.collect.android.logic.PropertyManager;
+import org.odk.collect.metadata.PropertyManager;
 import org.odk.collect.settings.keys.ProjectKeys;
 import org.odk.collect.shared.settings.Settings;
