+++ metadata/src/main/java/org/odk/collect/metadata/InstallIDProvider.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.metadata;
+package org.odk.collect.metadata;
 
 public interface InstallIDProvider {
     String getInstallID();
